Wrapper script for saving + plotting
